requirements.txt
einops==0.6.1
matplotlib==3.7.3
numpy==1.25.2
Pillow==10.0.0
pytorch-msssim==1.0.0
torch==2.0.1
torchvision==0.15.2